docs: Update readme file with greater detail
Added more detailed information to the readme file to help users understand the intention of this project, as well as how to install it and cloning the repo for editing in the Unity editor if they so desire.  Also converted from HTML to markdown to help learn how to do this for future repositories and readme edits.